Fix window title template CI guards
