+++ Sources/KeyboardShortcutSettingsFileStore.swift
@@ -392,7 +392,6 @@ final class CmuxSettingsFileStore {
 
         return snapshot
     }
-
     private func parseAppSection(
         _ section: [String: Any],
         sourcePath: String,
@@ -421,14 +420,10 @@ final class CmuxSettingsFileStore {
             }
             snapshot.managedUserDefaults[AppIconSettings.modeKey] = .string(mode.rawValue)
         }
-        if let raw = jsonString(section["windowTitleTemplate"]) {
-            snapshot.managedUserDefaults[WindowTitleTemplate.userDefaultsKey] = .string(raw)
-        } else if section.keys.contains("windowTitleTemplate") {
-            logInvalid("app.windowTitleTemplate", sourcePath: sourcePath)
-        }
         if let value = jsonBool(section["menuBarOnly"]) {
             snapshot.managedUserDefaults[MenuBarOnlySettings.menuBarOnlyKey] = .bool(value)
         }
+        if let raw = jsonString(section["windowTitleTemplate"]) { snapshot.managedUserDefaults[WindowTitleTemplate.userDefaultsKey] = .string(raw) } else if section.keys.contains("windowTitleTemplate") { logInvalid("app.windowTitleTemplate", sourcePath: sourcePath) }
         if let raw = jsonString(section["newWorkspacePlacement"]) {
             guard let placement = NewWorkspacePlacement(rawValue: raw) else {
                 logInvalid("app.newWorkspacePlacement", sourcePath: sourcePath)
