+++ web/messages/ja.json
@@ -824,7 +824,7 @@
         },
         "app": {
           "confirmQuit": "cmux の終了前に確認を求めるタイミングを制御します。DEV ビルドではこの設定に関係なく常にすぐ終了します。従来の app.warnBeforeQuit も boolean フォールバックとして引き続き受け付けます。",
-          "windowTitleTemplate": "任意の NSWindow タイトルテンプレートです。空欄のままなら、アクティブなワークスペース名を使う既定のタイトルを保ちます。対応するプレースホルダー: {windowId}, {windowToken}, {activeWorkspace}, {activeDirectory}, {defaultTitle}, {appName}。",
+          "windowTitleTemplate": "任意の NSWindow タイトルテンプレートです。空欄のままなら、アクティブなワークスペース名を使う既定のタイトルを保ちます。対応するプレースホルダー: '{windowId}', '{windowToken}', '{activeWorkspace}', '{activeDirectory}', '{defaultTitle}', '{appName}'。",
           "warnBeforeQuit": "終了確認用の従来の boolean フォールバックです。新しい設定では app.confirmQuit を使ってください。",
           "warnBeforeClosingTabXButton": "タブの閉じるボタンでタブを閉じる前に確認を表示します。",
           "hideTabCloseButton": "ペインのタブバーでタブの閉じるボタンを非表示にします。",
